test(toast): migrate layout
